refactor: replace datatypes.JSONMap for BambooScopeConfig.RepoMap
